refactor: github API 파일구조변경
