+++ src/api/index.ts
@@ -1,3 +1,8 @@
 //user
 export type { User } from '@/api/services/user/model'
 export { stackNames } from '@/api/services/user/model'
+
+export {
+  fetchGitHubContents,
+  fetchGitHubFile,
+} from '@/api/services/github/githubService'
